Move `junitMixin` to Specification DSL
