+++ src/test/java/specs/MockitoSpecJUnitStyle.java
@@ -10,6 +10,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 
 import com.greghaskins.spectrum.Spectrum;
+import com.greghaskins.spectrum.dsl.specification.Specification;
 
 import org.junit.Rule;
 import org.junit.runner.RunWith;
@@ -22,7 +23,7 @@ import org.mockito.junit.MockitoRule;
  * Demonstration of how to mix metaphors and use Mockito with Spectrum via
  * class members. There is only one instance of the test objects
  * so {@link org.mockito.InjectMocks} may have unexpected behaviour in some
- * complex situations. If this doesn't work use {@link Spectrum#junitMixin(Class)}.
+ * complex situations. If this doesn't work use {@link Specification#junitMixin(Class)}.
  */
 @RunWith(Spectrum.class)
 public class MockitoSpecJUnitStyle {
