+++ src/test/java/specs/MockitoSpecWithRuleClasses.java
@@ -1,10 +1,10 @@
 package specs;
 
 
-import static com.greghaskins.spectrum.Spectrum.junitMixin;
 import static com.greghaskins.spectrum.dsl.specification.Specification.beforeEach;
 import static com.greghaskins.spectrum.dsl.specification.Specification.describe;
 import static com.greghaskins.spectrum.dsl.specification.Specification.it;
+import static com.greghaskins.spectrum.dsl.specification.Specification.junitMixin;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.BDDMockito.given;
